+++ test/stream_test.py
@@ -2,17 +2,17 @@
 # -*- coding: utf-8 -*-
 
 import unittest
-import random
-import string
 
 import sys
 import os
+
+from techies import Queue, StateCounter
 
 target_path = os.path.join(os.path.dirname(__file__), '..', 'tidehunter')
 sys.path.append(target_path)
 
 # Test Targets
-from stream import StateCounter, Queue, Hunter
+from stream import Hunter
 
 
 class HunterTest(unittest.TestCase):
@@ -103,184 +103,6 @@ class HunterTestOAuth(HunterTest):
             'limit': 5,
             'timeout': 30
         }
-
-
-class QueueTest(unittest.TestCase):
-
-    def setUp(self):
-        self.key = 'test_q'
-        self.q = Queue(key=self.key, host='localhost', port=6379, db=0)
-        self.q.conn.delete(self.key)
-
-    def test_clear(self):
-        # When empty
-        self.q.clear()
-        self.assertEquals(self.q.conn.llen(self.key), 0)
-
-        # After putting
-        self.q.conn.rpush(self.key, 'test_val')
-        self.q.clear()
-        self.assertEquals(self.q.conn.llen(self.key), 0)
-
-    def test__len__(self):
-        # When empty
-        self.assertEqual(len(self.q), 0)
-
-        # When not empty
-        length = random.randint(1, 32)
-
-        for i in xrange(length):
-            self.q.conn.rpush(self.key, i)
-
-        self.assertEqual(len(self.q), length)
-
-    def test_full(self):
-        # Meaningless now
-        self.assertFalse(self.q.full())
-
-    def test_task_done(self):
-        self.q.task_done()
-
-    def test_join(self):
-        self.q.join()
-
-    def test_qsize(self):
-        # When empty
-        self.assertEqual(self.q.qsize(), 0)
-
-        # When not empty
-        length = random.randint(1, 32)
-
-        for i in xrange(length):
-            self.q.conn.rpush(self.key, i)
-
-        self.assertEqual(self.q.qsize(), length)
-
-    def test_empty(self):
-        # When empty
-        self.assertTrue(self.q.empty())
-
-        # When not empty
-        length = random.randint(1, 32)
-
-        for i in xrange(length):
-            self.q.conn.rpush(self.key, i)
-
-        self.assertFalse(self.q.empty())
-
-    def test_put(self):
-        val = ''.join(random.choice(string.ascii_uppercase + string.digits)
-                      for x in range(32))
-        self.assertTrue(self.q.put(val))
-        self.assertEqual(self.q.conn.lpop(self.key), val)
-
-    def test_get(self):
-        # When empty
-        # self.assertIsNone(self.q.get())
-        self.assertTrue(self.q.get() is None)  # 2.1 - 2.6 support
-
-        # When not empty
-        val = ''.join(random.choice(string.ascii_uppercase + string.digits)
-                      for x in range(32))
-        self.q.conn.rpush(self.key, val)
-        self.assertEqual(self.q.get(), val)
-
-    def test_put_nowait(self):
-        val = ''.join(random.choice(string.ascii_uppercase + string.digits)
-                      for x in range(32))
-        self.assertTrue(self.q.put_nowait(val))
-        self.assertEqual(self.q.conn.lpop(self.key), val)
-
-    def test_get_nowait(self):
-        # When empty
-        # self.assertIsNone(self.q.get())
-        self.assertTrue(self.q.get_nowait() is None)  # 2.1 - 2.6 support
-
-        # When not empty
-        val = ''.join(random.choice(string.ascii_uppercase + string.digits)
-                      for x in range(32))
-        self.q.conn.rpush(self.key, val)
-        self.assertEqual(self.q.get(), val)
-
-    def tearDown(self):
-        self.q.conn.delete(self.key)
-
-
-class StateCounterTest(unittest.TestCase):
-
-    def setUp(self):
-        self.key = 'test_sc'
-        self.sc = StateCounter(key=self.key, host='localhost', port=6379,
-                               db=0)
-
-    def test_clear(self):
-        # Right after init
-        self.sc.clear()
-        self.assertEqual(self.sc.conn.hget(self.key, 'count'), '0')
-
-        # After having some count
-        self.sc.conn.hincrby(self.key, 'count', 1)
-        self.sc.clear()
-        self.assertEqual(self.sc.conn.hget(self.key, 'count'), '0')
-
-    def test_str(self):
-        self.assertEqual(str(self.sc), 'State: 1\tCount: 0\tTotal: 0')
-
-    def test_get_state(self):
-        self.assertEqual(self.sc.get_state(), 1)
-
-    def test_get_count(self):
-        self.assertEqual(self.sc.get_count(), 0)
-
-    def test_get_total(self):
-        self.assertEqual(self.sc.get_total(), 0)
-
-    def test_get_all(self):
-        expected = {'count': '0', 'state': '1', 'total': '0'}
-        actual = self.sc.get_all()
-
-        self.assertEqual(actual, expected)
-
-    def test_start(self):
-        self.sc.conn.hset(self.key, 'state', 0)
-        self.assertEqual(self.sc.get_state(), 0)
-        self.sc.conn.hset(self.key, 'count', 100)
-        self.assertEqual(self.sc.get_count(), 100)
-
-        self.sc.start()
-
-        self.assertEqual(self.sc.get_state(), 1)
-        self.assertEqual(self.sc.get_count(), 0)
-
-    def test_stop(self):
-        self.assertEqual(self.sc.get_state(), 1)
-        self.sc.conn.hset(self.key, 'count', 100)
-        self.assertEqual(self.sc.get_count(), 100)
-        self.assertEqual(self.sc.get_total(), 0)
-
-        self.sc.stop()
-
-        self.assertEqual(self.sc.get_state(), 0)
-        self.assertEqual(self.sc.get_count(), 0)
-        self.assertEqual(self.sc.get_total(), 100)
-
-    def test_incr(self):
-        self.assertEqual(self.sc.get_count(), 0)
-        self.sc.incr()
-        self.assertEqual(self.sc.get_count(), 1)
-
-    def test_started(self):
-        self.assertTrue(self.sc.started())
-        self.sc.stop()
-        self.assertFalse(self.sc.started())
-
-    def test_stopped(self):
-        self.assertFalse(self.sc.stopped())
-        self.sc.stop()
-        self.assertTrue(self.sc.stopped())
-
-    def tearDown(self):
-        self.sc.conn.delete(self.key)
 
 
 if __name__ == '__main__':