+++ CHANGES.md
@@ -1,5 +1,10 @@
 ## Changelog
 
+### 0.2.0 (2014-01-22)
+
+* Moved codebase of `Queue`, `StateCounter` to [`techies`](https://github.com/woozyking/techies), which now becomes a dependency of `tidehunter`
+* Exposed accessibility of `Hunter` directly so you may use it as such, `from tidehunter import Hunter`
+
 ### 0.1.9 (2013-12-24)
 
 * `PyCurl` and `Redis` Python libraries bumped to the latest versions.